Add fiat currency storage management for desktop and extension clients
